requirements.txt
imgurpython==1.1.7
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
